agent/requirements.txt
anthropic>=0.18.0
web3>=6.15.0
python-dotenv>=1.0.0
langgraph>=0.0.20
langchain>=0.1.0
langchain-anthropic>=0.1.0
requests>=2.31.0
eth-account>=0.10.0
hexbytes>=0.3.0
